class Dom modules added fine
